[MIG] cooperator_worker: Migration to 16.0
