+++ cooperator_worker/__manifest__.py
@@ -4,7 +4,7 @@
     "author": "BEES coop, Coop IT Easy SC",
     "website": "https://github.com/beescoop/Obeesdoo",
     "category": "Cooperative management",
-    "version": "12.0.2.0.1",
+    "version": "16.0.1.0.0",
     "depends": [
         "shift",
         "cooperator",
