+++ cooperator_worker/views/product_template_views.xml
@@ -5,16 +5,18 @@
         <field name="model">product.template</field>
         <field name="inherit_id" ref="cooperator.product_template_form_view" />
         <field name="arch" type="xml">
-            <field name="customer" position="after">
-                <field
-                    name="allow_working"
-                    attrs="{'invisible':[('is_share','=',False)]}"
-                />
-                <field
-                    name="allow_shopping"
-                    attrs="{'invisible':[('is_share','=',False)]}"
-                />
-            </field>
+            <group name="configuration" position="after">
+                <group name="shifts" string="Shift">
+                    <field
+                        name="allow_working"
+                        attrs="{'invisible':[('is_share','=',False)]}"
+                    />
+                    <field
+                        name="allow_shopping"
+                        attrs="{'invisible':[('is_share','=',False)]}"
+                    />
+                </group>
+            </group>
         </field>
     </record>
 </odoo>
